+++ lib/screens/home/components/body.dart
@@ -4,6 +4,7 @@ import 'package:plant_app/constants.dart';
 import 'header_with_searchbox.dart';
 import 'title_with_more_btn.dart';
 import 'recomend_plants.dart';
+import 'featured_plants.dart';
 
 // import 'package:plant_app/constants.dart';
 // import 'package:flutter_svg/svg.dart';
@@ -28,50 +29,9 @@ class Body extends StatelessWidget {
           // 全幅の40％をカバーする
           const RecomendsPlants(),
           TitleWithMoreBtn(title: 'Featured Plants', press: () {}),
-          Row(
-            children: [
-              FeaturePlantCard(
-                image: 'assets/images/bottom_img_1.png',
-                press: () {},
-              ),
-            ],
-          ),
+          FeaturedPlants(),
+          const SizedBox(height: kDefaultPadding),
         ],
-      ),
-    );
-  }
-}
-
-class FeaturePlantCard extends StatelessWidget {
-  const FeaturePlantCard({
-    super.key,
-    required this.image,
-    required this.press,
-  });
-
-  final String image;
-  final void Function()? press;
-
-  @override
-  Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
-    return GestureDetector(
-      onTap: press,
-      child: Container(
-        margin: const EdgeInsets.only(
-          left: kDefaultPadding,
-          top: kDefaultPadding / 2,
-          bottom: kDefaultPadding / 2,
-        ),
-        width: size.width * 0.8,
-        height: 185,
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10),
-          image: DecorationImage(
-            fit: BoxFit.cover,
-            image: AssetImage(image),
-          ),
-        ),
       ),
     );
   }
